optimized scripts

+++ new_train.py
@@ -327,7 +327,7 @@ def train(cfg: DictConfig):
     )
 
     deepspeed_plugin = DeepSpeedPlugin(zero_stage=2, gradient_accumulation_steps=cfg.training.gradient_accumulation_steps, gradient_clipping=1.0)
-    accelerator = Accelerator(mixed_precision='bf16', deepspeed_plugin =deepspeed_plugin) #,  mixed_precision='fp16',
+    accelerator = Accelerator(mixed_precision='fp16', deepspeed_plugin =deepspeed_plugin) #,  mixed_precision='fp16',
     
     accelerator.wait_for_everyone()
     device= accelerator.device
